Services foi refatorada para evitar repetição e vazamento de código.

+++ NetMaster.Domain.csproj/Models/DataModels/StorageInfoModel.cs
@@ -3,8 +3,8 @@
     public class StorageInfoModel
     {
         public string? DeviceID { get; set; }
-        public int Size_GB { get; set; }
-        public int FreeSpace_GB { get; set; }
+        public double Size_GB { get; set; }
+        public double FreeSpace_GB { get; set; }
         public string? PSComputerName { get; set; }
         public string? Timestamp { get; set; }
         public string? IpAddress { get; set; }
